flaskr/requirements.txt
flask
flask-jwt-extended
flask-cors
mysql-connector-python
werkzeug
python-dotenv